Revert "Revert "Refactor flow diagram to use same structure as web app (#362)" (#364)"
This reverts commit bffcb73ee9fea369af9fd40b606dcdf02d34506a.
