+++ cli/daemon/dash/dashapp/src/components/FlowDiagram/FlowDiagram.tsx
@@ -1,6 +1,4 @@
 import React, { FC, useEffect, useState } from "react";
-import { APIMeta } from "~c/api/api";
-import JSONRPCConn, { NotificationMsg } from "~lib/client/jsonrpc";
 import {
   getEdgesFromMetaData,
   getNodesFromMetaData,
@@ -9,10 +7,9 @@ import {
   NodeData,
   PositionedNode,
   ServiceNode,
-} from "~lib/graph/graph-utils";
+} from "./flow-utils";
 import { ProvidedZoom } from "@visx/zoom/lib/types";
-import { getElkGraphLayoutData } from "~lib/graph/elk-algo";
-import { ProcessReload } from "~lib/client/client";
+import { getElkGraphLayoutData } from "./algorithms/elk-algo";
 import { Group } from "@visx/group";
 import { Zoom } from "@visx/zoom";
 import {
@@ -21,10 +18,11 @@ import {
   EncoreArrowHeadSVG,
   ServiceSVG,
   TopicSVG,
-} from "~c/svgElements";
+} from "./flowSvgElements";
 import { ParentSize } from "@visx/responsive";
+import { APIMeta } from "~c/api/api";
 
-class AppDiagramLayoutOptions implements GraphLayoutOptions {
+class LayoutOptions implements GraphLayoutOptions {
   serviceNodeMinWidth = 220;
   topicNodeMinWidth = 50;
   serviceNodeMinHeight = 60;
@@ -46,12 +44,10 @@ class AppDiagramLayoutOptions implements GraphLayoutOptions {
 }
 
 interface Props {
-  appID: string;
-  conn: JSONRPCConn;
+  metaData: APIMeta;
 }
 
-const AppDiagram: FC<Props> = ({ appID, conn }) => {
-  const [metaData, setMetaData] = useState<APIMeta>();
+const FlowDiagram: FC<Props> = ({ metaData }) => {
   const [graphLayoutData, setGraphLayoutData] = useState<GraphData>();
   const [activeNode, setActiveNode] = useState<PositionedNode | null>(null);
   const [activeDescendants, setActiveDescendants] = useState<string[] | null>(
@@ -62,7 +58,7 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
   const graphHeight = graphLayoutData?.height ?? 800;
 
   const getNumberOfEndpoints = (node: ServiceNode) => {
-    const svc = metaData!.svcs.find((s) => s.name === node.service_name);
+    const svc = metaData.svcs.find((s) => s.name === node.service_name);
     const endpoints = { public: 0, auth: 0, private: 0 };
     svc?.rpcs.forEach((rpc) => {
       if (rpc.access_type === "PUBLIC") endpoints.public++;
@@ -101,31 +97,11 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
   };
 
   useEffect(() => {
-    conn.request("status", { appID }).then((status: any) => {
-      if (status.meta) {
-        setMetaData(status.meta);
-      }
-    });
-    const onNotify = (msg: NotificationMsg) => {
-      if (msg.method === "process/reload") {
-        const data = msg.params as ProcessReload;
-        if (data.appID === appID) {
-          setMetaData(data.meta);
-        }
-      }
-    };
-    conn.on("notification", onNotify);
-    return () => {
-      conn.off("notification", onNotify);
-    };
-  }, []);
-
-  useEffect(() => {
     if (metaData) {
       getElkGraphLayoutData(
         getNodesFromMetaData(metaData),
         getEdgesFromMetaData(metaData),
-        new AppDiagramLayoutOptions()
+        new LayoutOptions()
       ).then(setGraphLayoutData);
     }
   }, [metaData]);
@@ -134,12 +110,12 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
 
   return (
     <div
-      className="relative flex w-full flex-col items-center justify-center"
-      style={{ height: "calc(100vh - 120px)", background: "#EEEEE1" }}
+      className="relative flex h-full w-full flex-col items-center justify-center"
+      style={{ background: "#EEEEE1" }}
     >
       {!graphLayoutData.nodes.length ? (
         <div>
-          <p>Add an encore service to your app and it will show up here</p>
+          <p>Add an service to your app and it will show up here</p>
         </div>
       ) : (
         <ParentSize>
@@ -162,8 +138,8 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
                 initialTransformMatrix={{
                   scaleX: 1,
                   scaleY: 1,
-                  translateX: parent.width / 2 - graphLayoutData?.width! / 2,
-                  translateY: parent.height / 2 - graphLayoutData?.height! / 2,
+                  translateX: parent.width / 2 - graphLayoutData.width! / 2,
+                  translateY: parent.height / 2 - graphLayoutData.height! / 2,
                   skewX: 0,
                   skewY: 0,
                 }}
@@ -171,7 +147,7 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
                 {(zoom) => (
                   <div className="relative">
                     <svg
-                      className="app-diagram"
+                      id="flow-diagram"
                       width={parent.width}
                       height={parent.height}
                       ref={zoom.containerRef}
@@ -269,4 +245,4 @@ const AppDiagram: FC<Props> = ({ appID, conn }) => {
   );
 };
 
-export default AppDiagram;
+export default FlowDiagram;
