+++ cli/daemon/dash/dashapp/src/components/FlowDiagram/algorithms/dagre-algo.ts
@@ -5,17 +5,13 @@ import {
   OUTSIDE_FACING_NODE_ID,
   PositionedEdge,
   PositionedNode,
-} from "~lib/graph/graph-utils";
+} from "../flow-utils";
 
 /**
  * Directed graph layout
  * https://github.com/dagrejs/dagre/wiki
  */
-export const getGraphLayoutData: GetGraphLayoutData = (
-  nodes,
-  edges,
-  options
-) => {
+export const getGraphLayoutData: GetGraphLayoutData = (nodes, edges, options) => {
   const graphData = new dagre.graphlib.Graph<NodeData>();
 
   graphData.setGraph({
@@ -42,10 +38,7 @@ export const getGraphLayoutData: GetGraphLayoutData = (
       width: options.getNodeWidth(nodeData),
       height: options.getNodeHeight(nodeData),
     });
-    if (
-      options.drawOutsideDependencyToNode &&
-      options.drawOutsideDependencyToNode(nodeData)
-    ) {
+    if (options.drawOutsideDependencyToNode && options.drawOutsideDependencyToNode(nodeData)) {
       graphData.setEdge(OUTSIDE_FACING_NODE_ID, nodeData.id, { weight: 10000 });
     }
   });
@@ -56,9 +49,7 @@ export const getGraphLayoutData: GetGraphLayoutData = (
 
   dagre.layout(graphData, {});
 
-  const positionNodes: PositionedNode[] = graphData
-    .nodes()
-    .map((n) => graphData.node(n));
+  const positionNodes: PositionedNode[] = graphData.nodes().map((n) => graphData.node(n));
   const positionedEdges: PositionedEdge[] = graphData
     .edges()
     .map((e) => graphData.edge(e) as PositionedEdge);
