+++ cli/daemon/dash/dashapp/src/main.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import "./app.css";
+import "./components/FlowDiagram/flow.css";
 import App from "./App";
 import { createRoot } from "react-dom/client";
